expose only function to import account data on client
By doing this and moving the file handling to the CLI we avoid future issues if we need to run this with no_std. And by keeping the import account data function we can still test with a test store instead of overwriting the dev ones
